Change eventId_t from int8_t to int16_t in case if more than 127 events are needed

+++ OnPod/Node/Common/EventScheduler/include/event.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-typedef int8_t eventId_t; // Should range from 0-127. -1 indicates error in finding free ID
+typedef int16_t eventId_t; // Should range from 0-127. -1 indicates error in finding free ID
 typedef uint32_t timeMs_t; // Time in ms
 
 enum class EventType
